task-2/requirements.txt
faiss-cpu
sentence-transformers
numpy 
pandas 
spacy 